+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 #### Fixes
 
 * Make `AngularSnapshotSerializer` compatible with Ivy ([#366](https://github.com/thymikee/jest-preset-angular/pull/366))
+* Make `AngularSnapshotSerializer` remove auto-generated angular classes in addition to attributes
 
 ### v8.1.2